+++ heylogs-api/src/test/java/internal/heylogs/ExtendedRulesTest.java
@@ -65,5 +65,8 @@ public class ExtendedRulesTest {
     public void testValidateConsistentSeparator() {
         assertThat(validateConsistentSeparator(using("/ErraticSeparator.md")))
                 .isEqualTo(Failure.builder().rule(CONSISTENT_SEPARATOR).message("Expecting consistent version-date separator \\u002d, found [\\u002d, \\u2013, \\u2014]").line(1).column(1).build());
+
+        assertThat(validateConsistentSeparator(using("/NonDefaultSeparator.md")))
+                .isEqualTo(NO_PROBLEM);
     }
 }